Refactor to move logger context to middlewares package.
